backend/requirements.txt
fastapi==0.111.0
uvicorn==0.29.0
sqlalchemy==2.0.30
bcrypt==4.1.3
python-jose[cryptography]==3.3.0
python-multipart==0.0.9
openai==1.30.1
pydantic==2.7.1
aiosqlite==0.20.0
greenlet==3.0.3
python-dotenv==1.0.1
